reimplement qChecksum() to calculate CRC-32 checksum
nearly 2x faster, benchmark result:
***************************************************
 Tag                   * Ticks CRC-32/CRC-16 (Winner) *
***************************************************
 bench_qchecksum_notag : 1947841/26951171 (CRC-32)

***************************************************
 Total                 : 1947841/26951171 (CRC-32)
***************************************************

+++ src/core/tools/qbytearray.h
@@ -68,8 +68,7 @@ inline int qstrnicmp(const char *str1, const char *str2, uint len)
         : (str1 ? 1 : (str2 ? -1 : 0));
 }
 
-// qChecksum: Internet checksum
-Q_CORE_EXPORT quint16 qChecksum(const char *s, uint len);
+Q_CORE_EXPORT quint32 qChecksum(const char *s, uint len);
 
 Q_CORE_EXPORT QByteArray qRandomUuid();
